==> ECMO stroke/1_descriptives_ECMO.Rmd <==
---
title: "CCCC: V1 ECMO and stroke patients"
author: "Adrian Barnett, Lan Kelly"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  word_document: 
  reference_docx: rmarkdown-styles-SAP.docx
---
  
```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, 
error=FALSE, comment='', dpi=400)
options(knitr.kable.NA = '')
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
source = 'real_data'
#source = 'dummy_data'
source('1_ecmo_ich_data_prep.R') # prepares the data and runs exclusions
library(visdat)
library(summarytools)
library(pander)
library(naniar)
library(arsenal)

fup=30
```

This report uses the data from `r data_date`. Data is from patients who received 
ECMO at anytime during their ICU stay.


# Descriptive statistics

### Numbers on ECMO
Proportion of patients on any type of ECMO.

```{r}
tab_any = with(patients, freq(any_ecmo, round.digits=1, cumul = FALSE, report.nas	=FALSE))

# percent and 95% CI
r = sum(patients$any_ecmo=='Yes', na.rm=T)
n = nrow(patients)
p = r/n
z = qnorm(0.975)
se = sqrt(p*(1-p)/n)
lower = p - (z*se)
upper = p + (z*se)
```
```{r,  results='asis'}
knitr::kable(tab_any, digits=1)
```

The percent of patients experiencing any ECMO is `r roundz(p*100,1)`% with a 95% confidence interval from `r roundz(lower*100,1)`% to `r roundz(upper*100,1)`%.

\newpage
# Summary statistics for patients on ECMO

For the `r nrow(ecmo_patients)` patients on any type of ECMO.

## Demographics

### Continuous variables

Height is in cm and weight in kg.

```{r}
cont_vars = c('age','bmi' ,'height','weight','sofa','apache_ii')
#
tab = select(ecmo_patients, pin, all_of(cont_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin) %>%
  group_by(Variable) %>%
  summarise(Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep=''),
            Min = roundz(min(Value, na.rm=TRUE),0),
            Max = roundz(max(Value, na.rm=TRUE),0)) %>%
  select(Variable, Missing, Median, IQR, Min, Max) %>%
  mutate(Variable = case_when(Variable=='age' ~ 'Age', 
                              Variable == 'bmi' ~ 'BMI',
                              Variable == 'height' ~ 'Height',
                              Variable == 'weight' ~ 'Weight',
                              Variable == 'sofa' ~ 'SOFA',
                              Variable == 'apache_ii' ~ 'APACHE II'
                           ))
ftab = flextable(tab) %>% 
  theme_box() %>%
  autofit()
ftab
```

\newpage
### Categorical variables

```{r}
#
cat_vars = c('sex','ethnicity','country','ecmo_type','comorbidity_chronic_kidney_disease',
             'comorbidity_severe_liver_disease','comorbidity_diabetes','comorbidity_hypertension')
#
tab = select(ecmo_patients, pin, all_of(cat_vars)) %>%
  mutate_if(~is.character(.x), ~ifelse(.x=='', NA, .x)) %>%
  tidyr::gather(key='Variable', value='Value', -pin) %>%
  group_by(Variable, Value) %>%
  summarise(Count = n()) %>%
  group_by(Variable) %>%
  mutate(Percent = roundz(prop.table(Count)*100,0)) %>% # calculate percent
  arrange(Variable, -Count) %>% # arrange from high to low within variables
  mutate(Variable = case_when(
    Variable == 'sex' ~ 'Sex',
    Variable == 'ethnicity' ~ 'Ethnicity',
    Variable == 'country' ~ 'Country',
    Variable == 'ecmo_type' ~ 'ECMO Type',
    Variable == 'comorbidity_chronic_kidney_disease' ~ 'CKD',
    Variable == 'comorbidity_severe_liver_disease' ~ 'Severe Liver Disease',
    Variable == 'comorbidity_diabetes' ~ 'Diabetes',
    Variable == 'comorbidity_hypertension' ~ 'Hypertension'
  )) %>%
  ungroup() 
ftab = flextable(tab) %>%
  merge_v(j = c("Variable")) %>% # merge rows
  theme_box() %>%
  autofit()
ftab
```

The table rows are ordered by frequency.

\newpage
# Venous-Venous ECMO patients only
```{r}

n_any_ecmo = nrow(ecmo_patients)
ecmo_patients <- ecmo_patients %>%
  filter(ecmo_type == "Venous-Venous")

n_VV_episode=nrow(ecmo_patients)
n_VV_repeats = n_VV_episode - length(unique(ecmo_patients$pin))

```

The following analysis is for the `r nrow(ecmo_patients)` V-V ECMO patients only. `r n_VV_repeats` patients had more than one episode of ECMO.

\newpage
### Summaries for time variables in days

```{r}
time_vars = c('days_ecmo','days_hosp','days_vent','days_sympt_hosp','days_vent_ecmo','day')
# ecmo_patients <- ecmo_patients %>%
#   mutate(
#     days_ecmo = as.numeric(date_ecmo_discontinued - date_ecmo + 0.5),
#     days_hosp = as.numeric(date_hospital_discharge - date_admission + 0.5),
#     days_vent = as.numeric(date_mech_vent_discontinued - date_mechanical_ventilation + 0.5),
#     days_sympt_hosp = as.numeric(date_admission - date_first_symptom + 0.5),
#     day = as.numeric(date_ecmo - date_icu))
#
tab = select(ecmo_patients, pin, all_of(time_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin) %>%
  group_by(Variable) %>%
  mutate(Value = ifelse(Value < 0, NA, Value)) %>% # blank obvious errors
  summarise(Missing = sum(is.na(Value)),
            Median = roundz(median(Value, na.rm=TRUE),0),
            Q1 = roundz(quantile(Value, probs=0.25, na.rm=TRUE),0),
            Q3 = roundz(quantile(Value, probs=0.75, na.rm=TRUE),0),
            IQR = paste(Q1 , ' to ', Q3, sep=''),
            Min = roundz(min(Value, na.rm=TRUE),0),
            Max = roundz(max(Value, na.rm=TRUE),0)) %>%
  select(Variable, Missing, Median, IQR, Min, Max) %>%
  mutate(Variable = case_when(
    Variable == 'days_ecmo' ~ 'Length of the ECMO run',
    Variable == 'days_hosp' ~ 'Hospital length of stay',
    Variable == 'days_vent' ~ "Duration of mechanical ventilation",
    Variable == 'days_mech_vent' ~ "Days from mechanical ventilation to ECMO",
    Variable == 'days_sympt_hosp' ~ "Days from first symptoms to hospitalisation",
    Variable == 'day' ~ 'Day put on ECMO'
  ))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```


Just for the `r nrow(ecmo_patients)` on V-V ECMO. IQR is the inter-quartile range.


\newpage
## Other ECMO summaries 

```{r}
#
cat_vars = c('ecmo_drainage_cannula_insertion_site', 'ecmo_return_cannula_insertion_site',
             'cannula_lumen','ecmo_cardiac_arrest_before', 'ecmo_continuous_rpt_before')
#
tab = select(ecmo_patients, pin, all_of(cat_vars)) %>%
  tidyr::gather(key='Variable', value='Value', -pin) %>%
  group_by(Variable, Value) %>%
  summarise(N = n()) %>%
  group_by(Variable) %>%
  mutate(Percent = roundz(prop.table(N)*100,0)) %>% # calculate percent
  arrange(Variable, -N) %>% # arrange from high to low within variables
  ungroup() %>%
  mutate(
    # sort out missing
    Value = ifelse(is.na(Value)==TRUE, 'Missing', Value),
    Value = ifelse(Value=='', 'Missing', Value),
    # nicer labels
    Variable = ifelse(Variable=='ecmo_drainage_cannula_insertion_site', 'Drainage cannula insertion site', Variable),
    Variable = ifelse(Variable=='ecmo_return_cannula_insertion_site', 'Return cannula insertion site', Variable),
    
    Variable = ifelse(Variable=='cannula_lumen', 'Type of Cannula', Variable),
    Variable = ifelse(Variable=='ecmo_cardiac_arrest_before', 'Cardiac arrest before ECMO', Variable),
    Variable = ifelse(Variable=='ecmo_continuous_rpt_before', 'Continuous Renal Replacement Therapy before ECMO', Variable))
ftab = flextable(tab) %>%
  merge_v(j = c("Variable")) %>% # merge rows
  theme_box() %>%
  autofit()
ftab
```


Just for the `r nrow(ecmo_patients)` on V-V ECMO. 

\newpage
### Blood gases before and after ECMO

#### a) PaO2

Values of PaO2 < 40 after ECMO cannulation have been removed (assuming data entry error)

```{r, results='asis'}
pa_o2_data = blood_gases(before_var='ecmo_worst_pa_o2_6hr_before',
                         after_var = 'pa_o2',
                          etype="Venous-Venous") # from 99_functions.R
lplot = ggplot(data=pa_o2_data$to_plot, aes(x=xaxis, y=result, group=pin))+
  geom_line(col=grey(0.4))+
  geom_line(data=pa_o2_data$av_line, aes(x=xaxis, y=result, group=pin), col='dark red')+ # add average
  theme_bw()+
  scale_x_discrete(expand = c(0.05,0.05)) + # reduce white space
  xlab('')+
  ylab('PaO2 (log scale)')+
  scale_y_log10()
lplot
```

Here we examine the trends in PaO2 from 6 hours before V-V ECMO initiation to the first PaO2 in the period after ECMO initiation. 

The line plot shows the individual change in PaO2 for `r length(unique(pa_o2_data$to_plot$pin))` patients with available blood gas data and ECMO dates. The y-axis is on a log-scale because of the positive skew in PaO2. The red lines shows the average difference between the two times.

The mean percentage change between the worst before reading and first reading after ECMO was `r roundz(pa_o2_data$test$pdiff, 1)` with a 95% confidence interval from `r roundz(pa_o2_data$test$conf.int[1], 1)` to `r roundz(pa_o2_data$test$conf.int[2], 1)` and p-value `r format.pval(pa_o2_data$test$p.value, eps=0.001, digits=3)` (using a paired t-test with log-transformed data).

\newpage
#### b) PaCO2

Values of PaCO2 < 10 before ECMO cannulation have been removed (assuming data entry error)

```{r, results='asis'}
pa_co2_data = blood_gases(
                          before_var='ecmo_worst_pa_co2_6hr_before',
                          after_var = 'pa_co2',
                          etype="Venous-Venous") # from 99_functions.R
lplot = ggplot(pa_co2_data$to_plot, aes(x=xaxis, y=result, group=pin))+
  geom_line(col=grey(0.4))+
  geom_line(data=pa_co2_data$av_line, aes(x=xaxis, y=result, group=pin), col='dark red')+ # add average
  scale_x_discrete(expand = c(0.05,0.05)) + # reduce white space
  theme_bw()+
  xlab('')+
  ylab('PaCO2 (log-scale)')+
  scale_y_log10()
lplot
```

Here we examine the trends in PaO2 from 6 hours before V-V ECMO initiation to the first PaCO2 in the period after ECMO initiation.

The line plot shows the individual change in PaCO2 for `r length(unique(pa_o2_data$to_plot$pin))` patients with available blood gas data and ECMO dates. The y-axis is on a log-scale because of the positive skew in PaCO2.

The mean percentage change between the worst before reading and first reading after ECMO was `r roundz(pa_co2_data$test$pdiff, 1)` with a 95% confidence interval from `r roundz(pa_co2_data$test$conf.int[1], 1)` to `r roundz(pa_co2_data$test$conf.int[2], 1)` and p-value `r format.pval(pa_co2_data$test$p.value, eps=0.001, digits=3)` (using a paired t-test with log-transformed data).


\newpage
#### c) FiO2

```{r, results='asis'}
fi_o2_data = blood_gases(
                          before_var='ecmo_highest_fi_o2_6hr_before',
                          after_var = 'fi_o2',
                          etype="Venous-Venous") # from 99_functions.R

lplot =  ggplot(fi_o2_data$to_plot, aes(x=xaxis, y=result, group=pin))+
  geom_line(col=grey(0.4))+
  geom_line(data=fi_o2_data$av_line, aes(x=xaxis, y=result, group=pin), col='dark red')+ # add average
  scale_x_discrete(expand = c(0.05,0.05)) + # reduce white space
  theme_bw()+
  xlab('')+
  ylab('FiO2 (log-scale)')+
  scale_y_log10()
lplot
```

Here we examine the trends in FiO2 from 6 hours before V-V ECMO initiation to the first FiO2 in the period after ECMO initiation.

The line plot shows the individual change in FiO2 for `r length(unique(fi_o2_data$to_plot$pin))` patients with available blood gas data and ECMO dates. The y-axis is on a log-scale because of the positive skew in FiO2.

The mean percentage change between the worst before reading and first reading after ECMO was `r roundz(fi_o2_data$test$pdiff, 1)` with a 95% confidence interval from `r roundz(fi_o2_data$test$conf.int[1], 1)` to `r roundz(fi_o2_data$test$conf.int[2], 1)` and p-value `r format.pval(fi_o2_data$test$p.value, eps=0.001, digits=3)` (using a paired t-test with log-transformed data).

\newpage
#### d) PEEP


```{r, results='asis'}
peep_data = blood_gases(
                          before_var='ecmo_highest_peep_6hr_before',
                          after_var = 'eotd_peep',
                          etype="Venous-Venous",
                          logt=F) # from 99_functions.R
lplot =  ggplot(peep_data$to_plot, aes(x=xaxis, y=result, group=pin))+
  geom_line(col=grey(0.4))+
  geom_line(data=peep_data$av_line, aes(x=xaxis, y=result, group=pin), col='dark red')+ # add average
  scale_x_discrete(expand = c(0.05,0.05)) + # reduce white space
  theme_bw()+
  xlab('')+
  ylab('PEEP') #+
  #scale_y_log10()
lplot
```

Here we examine the trends in PEEP from 6 hours before V-V ECMO initiation to the first PEEP in the period after ECMO initiation.

The line plot shows the individual change in PEEP for `r length(unique(peep_data$to_plot$pin))` patients with available blood gas data and ECMO dates. The y-axis is on a linear scale.

The mean percentage change between the worst before reading and first reading after ECMO was `r roundz(peep_data$test$pdiff, 2)` with a 95% confidence interval from `r roundz(peep_data$test$conf.int[1], 2)` to `r roundz(peep_data$test$conf.int[2], 2)` and p-value `r format.pval(peep_data$test$p.value, eps=0.001, digits=2)` (using a paired t-test).


\newpage

# Values collected during the ICU stay

```{r, include=FALSE}
# data to plot, must have dates
combined_ecmo <- combined_ecmo %>% filter(ecmo_type == "Venous-Venous")

to_plot = filter(combined_ecmo, 
                 !is.na(date_icu),
                 !is.na(date_daily)) %>%
  mutate(day = as.numeric(date_daily - date_icu)) %>%
  filter(day <= 30) # exclude long stays for the plot
```

The plots below are limited to the first `r fup` days during the ICU stay for V-V ECMO patients only. `r combined_ecmo %>% filter(is.na(date_icu) | is.na(date_daily)) %>% select(pin) %>% unique %>% nrow()` patients with missing or implausible dates have been excluded.

\newpage
## Platelet count

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=platelet_count))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('Platelet count')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>%
  summarise(mean = mean(platelet_count, na.rm = TRUE),
            se = sem(platelet_count)) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
         #day = ifelse(any_ecmo=='Yes', day+0.2, day)
         ) # jitter x slightly to avoid overlap
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('Platelet count')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO', ,'cyan4'
  g.theme
tplot
```

Just for the first `r fup` days and the `r length(unique(to_plot$pin))` patients in ICU. 

\newpage
## Blood urea nitrogen

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=blood_urea_nitrogen))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('Blood urea nitrogen')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO', ,'cyan4'
  g.theme
tplot
```


##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>% #, any_ecmo
  summarise(mean = mean(blood_urea_nitrogen, na.rm = TRUE),
            se = sem(blood_urea_nitrogen)) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
         #day = ifelse(any_ecmo=='Yes', day+0.2, day)
         ) # jitter x slightly to avoid overlap
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('Blood urea nitrogen')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

Just for the first 30 days and the `r length(unique(to_plot$pin))` patients in ICU.

\newpage
## Bilirubin

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=bilirubin))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('Bilirubin')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```


##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>%  #, any_ecmo
  summarise(mean = mean(bilirubin, na.rm = TRUE),
            se = sem(bilirubin)) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
        # day = ifelse(any_ecmo=='Yes', day+0.2, day)
         ) # jitter x slightly to avoid overlap
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('Bilirubin')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

Just for the first `r fup` days and the `r length(unique(to_plot$pin))` patients in ICU.

\newpage
## APTT/APTR

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=aptt_aptr))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('APTT/APTR')+
  scale_color_manual(values=c('darkorange1'))+ #'Any ECMO', ,'cyan4'
  g.theme
tplot
```

##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>%  #, any_ecmo
  summarise(mean = mean(aptt_aptr, na.rm = TRUE),
            se = sem(aptt_aptr)) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
         #day = ifelse(any_ecmo=='Yes', day+0.2, day) # jitter x slightly to avoid overlap
         ) 
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('APTT/APTR')+
  scale_color_manual( values=c('darkorange1'))+  #'Any ECMO',,'cyan4'
  g.theme
tplot
```

Just for the first 30 days and the `r length(unique(to_plot$pin))` patients in ICU.

\newpage
## INR

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=inr))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('INR')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```


##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>%  #, any_ecmo
  summarise(mean = mean(inr, na.rm = TRUE),
            se = sem(inr)) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
         #day = ifelse(any_ecmo=='Yes', day+0.2, day) # jitter x slightly to avoid overlap
         ) 
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('INR')+
  scale_color_manual(values=c('darkorange1'))+ #'Any ECMO', ,'cyan4'
  g.theme
tplot
```

Just for first 30 days and the `r length(unique(to_plot$pin))` patients in ICU.

\newpage
## ALT-SGPT

For this variable I use a log-transform because the variable has a very strong positive skew.

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=alt_sgpt))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('ALT-SGPT (log-scale)')+
  scale_y_log10()+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

The y-axis is on a log scale.

##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day, any_ecmo) %>%
  summarise(mean = mean(log(alt_sgpt), na.rm = TRUE),
            se = sem(log(alt_sgpt))) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se),
         day = ifelse(any_ecmo=='Yes', day+0.2, day)) # jitter x slightly to avoid overlap
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('Log-transformed ALT-SGPT')+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

Just for the first `r fup` days and the `r length(unique(to_plot$pin))` patients in ICU.

The y-axis is for the log-transformed value.

\newpage
## AST-SGOT

For this variable I use a log-transform because the variable has a very strong positive skew.

##### Boxplot over time

```{r, fig.width=8}
tplot = ggplot(data=to_plot, aes(x=factor(day), y=ast_sgot))+ #, col=factor(any_ecmo)
  geom_boxplot()+
  xlab('Day in ICU')+
  ylab('AST-SGOT (log-scale)')+
  scale_y_log10()+
  scale_color_manual( values=c('darkorange1'))+ #'Any ECMO',,'cyan4'
  g.theme
tplot
```

The y-axis is on a log scale.

##### Plot of means and 95% confidence intervals for the means over time

```{r, fig.width=8}
to_stat = group_by(to_plot, day) %>% #, any_ecmo
  summarise(mean = mean(log(ast_sgot), na.rm = TRUE),
            se = sem(log(ast_sgot))) %>%
  ungroup() %>%
  mutate(lower = mean - (z*se),
         upper = mean + (z*se)
         #,day = ifelse(any_ecmo=='Yes', day+0.2, day) # jitter x slightly to avoid overlap
         ) 
tplot = ggplot(data=to_stat, aes(x=day, y=mean, ymin=lower, ymax=upper))+ #, col=factor(any_ecmo)
  geom_point()+
  geom_errorbar(width=0)+
  xlab('Day in ICU')+
  ylab('Log-transformed AST-SGOT')+
  scale_color_manual(values=c('darkorange1'))+ #'Any ECMO', ,'cyan4'
  g.theme
tplot
```

Just for the first `r fup` days and the `r length(unique(to_plot$pin))` patients in ICU.

The y-axis is for the log-transformed value.

\newpage
## Anticoagulants in the last 24 hours

```{r, fig.width=8}
to_plot = mutate(to_plot, facet = paste("Any ECMO = ", any_ecmo, sep=''), # nice facet
                 eotd_anticoagulants_plot = ifelse(is.na(eotd_anticoagulants), 2, eotd_anticoagulants)) # replace missing %>%
bplot = ggplot(data=to_plot, aes(x=factor(day), fill=factor(eotd_anticoagulants_plot)))+
  geom_bar(position='stack')+
  xlab('Day in ICU')+
  ylab('Number of patients')+
  scale_fill_manual('Anticoagulants', values=c('indianred1','cyan3','grey'), labels=c('No','Yes','Missing'))+
  g.theme+
  theme(axis.text.x = element_text(size=5))#+ # reduce label size
  #facet_wrap(~facet, scales='free_y')
bplot
```


Same as the previous plot, broken down by whether patients had a stroke or not.

```{r, fig.width=8}
to_plot = mutate(to_plot, facet = paste("Stroke = ", complication_stroke, sep=''), # nice facet
                 eotd_anticoagulants_plot = ifelse(is.na(eotd_anticoagulants), 2, eotd_anticoagulants)) # replace missing %>%
bplot = ggplot(data=to_plot %>% 
                 filter(!is.na(complication_stroke)), 
               aes(x=day, fill=factor(eotd_anticoagulants_plot)))+
  geom_bar(position='stack')+
  xlab('Day in ICU')+
  ylab('Number of patients')+
  scale_fill_manual('Anticoagulants', values=c('indianred1','cyan3','grey'), labels=c('No','Yes','Missing'))+
  g.theme+
  theme(axis.text.x = element_text(size=5)) + # reduce label size
  facet_wrap(~facet, scales='free_y')
bplot
```

\newpage
## Anticoagulants type in the last 24 hours

```{r, fig.width=8}
to_plot_anti = filter(to_plot, eotd_anticoagulants_type!='')
colours = gg_color_hue(10)
bplot = ggplot(data=to_plot_anti, aes(x=factor(day), fill=factor(eotd_anticoagulants_type)))+
  geom_bar(position='stack')+
  xlab('Day in ICU')+
  ylab('Number of patients')+
  scale_fill_manual(NULL, values=colours)+
  g.theme+
  theme(axis.text.x = element_text(size=5))#+ # reduce label size
 # facet_wrap(~facet, scales='free_y')
bplot
```

\newpage
## Haemorrhagic complication

```{r, fig.width=8}
to_plot = mutate(to_plot, eotd_haemorrhagic_complication_plot = ifelse(is.na(eotd_haemorrhagic_complication), 2, eotd_haemorrhagic_complication)) # make extra category for missing
bplot = ggplot(data=to_plot, aes(x=factor(day), fill=factor(eotd_haemorrhagic_complication_plot)))+
  geom_bar(position='stack')+
  xlab('Day in ICU')+
  ylab('Number of patients')+
  scale_fill_manual('Haemorrhagic\ncomplication', values=c('indianred1','cyan3','grey'), labels=c('No','Yes','Missing'))+
  g.theme+
  theme(axis.text.x = element_text(size=5)) #+ # reduce label size
  #facet_wrap(~facet, scales='free_y')
bplot
```

\newpage
## Source of Haemorrhagic Complication 

```{r, results='asis'}
to_table = filter(to_plot, eotd_haemorrhagic_complication==1)
knitr::kable(freq(to_table$eotd_haemorrhagic_complication_source, cumul=FALSE, report.nas = FALSE), digits=1)
```

Just for the `r nrow(to_table)` days with a haemorrhagic complication in the first `r fup` days. The total number of patients is `r length(unique(to_table$pin))`.

\newpage
### Number of patients with "central nervous system"

```{r, include=FALSE}
to_count = filter(to_plot, eotd_haemorrhagic_complication==1,
                  eotd_haemorrhagic_complication_source == 'Central nervous system'
) %>%
  select(pin, complication_stroke) %>%
  unique()
n_patients_cns = nrow(to_count)
```

The previous table is the number of days and so contains repeated days from the same patients. The number of unique patients with at least one day of "central nervous system" is `r n_patients_cns`.

The table below shows whether these `r n_patients_cns` patients also had stroke (using the complication data).

```{r, results='asis'}
knitr::kable(freq(to_count$complication_stroke, cumul=FALSE, report.nas = TRUE), digits=1)
```

\newpage
# Neurological Complications in ECMO patients

## Number of strokes

All V-V ECMO patients.  A blank cell means no data was entered.

```{r, results='asis'}
tab_stroke = with(ecmo_patients, freq(complication_stroke, round.digits=0, report.nas	=FALSE, cumul = FALSE))
knitr::kable(tab_stroke, digits=1)
```


The next table shows whether the stroke was after initiation of ECMO.
```{r, results='asis'}
tab_stroke2 = with(ecmo_patients %>% filter(complication_stroke == "Yes"), freq(stroke_after_ecmo, round.digits=0, report.nas	=FALSE, cumul = FALSE))
knitr::kable(tab_stroke2, digits=1)


```

### Stroke by anticoagulant use

This table is a cross-tabulation of anticoagulant use by stroke for V-V ECMO patients, **excluding those with stroke before ECMO**. A blank cell is missing data.

```{r, results='asis'}
to_stat = combined_ecmo %>% 
  filter(ecmo_type == "Venous-Venous" 
         #,is.na(stroke_before_ecmo) | stroke_before_ecmo != "Yes"
         ) %>%
  group_by(pin) %>%
  slice_head(n=1) %>%
  ungroup() %>%
  select(pin, anticoagulants, complication_stroke, stroke_before_ecmo) %>%
  mutate(complication_stroke = ifelse(is.na(complication_stroke), "Missing",
                                            complication_stroke))

labels(to_stat) <- c(anticoagulants = "Anticoagulants",
       complication_stroke = "Stroke")
tab_ac <- tableby(complication_stroke ~ notest(anticoagulants), data=to_stat,
                cat.stats=c("Nmiss","countpct"),
                na.action=na.tableby(T),
                 test=FALSE, total=T)
summary(tab_ac)

# tab_anticoag = with(to_stat, ctable(anticoagulants, complication_stroke,  chisq=FALSE,
#                                       prop='c', round.digits=0, dnn=c(' ','.')))
# tab_anticoag
```
 


\newpage
## Stroke types


```{r}
# remove unknown and empty (March 2021)
to_plot = ecmo_patients %>% filter(stroke_after_ecmo == "Yes"  )

#Patients with no stroke type have been excluded from this plot.
# to_remove = NULL
# for (k in 1:nrow(to_plot)){ # had to loop because of list
#   if(is.null(to_plot$stroke_type[[k]]) == TRUE){to_remove = c(to_remove, k)}
#   index = to_plot$stroke_type[[k]] %in% c('Other','Unknown')
#   if(any(index) == TRUE){
#     to_plot$stroke_type[[k]] = 'Unspecified'
#   }
# }
# to_plot = to_plot[1:nrow(to_plot) %in% to_remove==FALSE, ]
# # add ggupset
# cplot = ggplot(data=to_plot, aes(x = stroke_type)) +
#   geom_bar(aes(y=..count..), fill='seagreen') +
#   theme_bw() +
#   xlab("") +
#   ylab("Number of patients") +
#   scale_x_upset( )+
#   scale_y_continuous(breaks=seq(0,20,2))+ # avoid fractions in breaks
#   g.theme +
#   theme(plot.margin = margin(t = 5, r = 0, b = 0, l = 110, unit = "pt")) #110
#cplot
#jpeg('figures/stroke_upset.jpg', width=5, height=4.5, units='in', res=300, quality=100)
#print(cplot)
#invisible(dev.off())


# Up to two options for the stroke type variable could be ticked.  The plot is just for the `r nrow(to_plot)` V-V ECMO patients with a stroke complication **after** initiation of ECMO.
# 
# The most common response with complete data is "Intraparenchymal haemorrhage". The second
# most common is "Subarachnoid haemorrhage".
```



### Combined stroke types

Intraparenchymal and Subarachnoid  haemorrhage have been combined into Hemorrhage while Ischemic stroke and Hypoxic ischemic brain injury have been combined into "Ischemic".

For the `r nrow(to_plot)` V-V ECMO patients with a stroke complication **after** initiation of ECMO.

```{r, results='asis'}
tab_stroke2 = with(to_plot, freq(stroke_group, round.digits=0, report.nas	=FALSE, cumul = FALSE))
knitr::kable(tab_stroke2, digits=1)


```

\newpage
## Other Neurological complications

### Seizures

```{r, results='asis'}
tab_stroke = with(ecmo_patients, freq(complication_seizure, round.digits=0, report.nas	=FALSE, cumul = FALSE))
knitr::kable(tab_stroke, digits=1)
```


\newpage
### Neurological complications recorded as free text

Other neurological complications other than stroke are included in the text entries in the table below. These complications can be included by looking for keywords in the list eg “neuropathy”, “encephalopathy”. The following table is for patients **WITHOUT a recorded complication of stroke**.

```{r, results='asis'}
# tab_stroke = with(ecmo_patients %>% filter(complication_stroke == "No"), 
#                   freq(complication_other_value, round.digits=0, 
#                                       report.nas	=FALSE, cumul = FALSE))
# knitr::kable(tab_stroke, digits=1)



summary(tableby(~complication_other_value, data = ecmo_patients %>% filter(complication_stroke == "No")))

```

Patients **without** a recorded complication of stroke, who may have had a stroke, are shown in the following table.

```{r, results='asis'}
strokelist <- c("ICH with midline shift","Intracerebral bleeding",
                 "intracerebral hemorrhage, subarachnoid hemorrhage,·")
p<- patients %>% 
  filter(trimws(complication_other_value) %in% strokelist | 
           grepl("cerebral", complication_other_value, ignore.case=T),
         complication_stroke != "Yes") %>% 
  select(pin, complication_stroke, stroke_during_treatment, complication_other_value)

knitr::kable(p)

```

 
\newpage
## Association between a cannula in the internal jugular vein and stroke/ICH/death

The table below shows the association between insertion site and stroke complication.

```{r, results='asis'}
tab = group_by(ecmo_patients, ecmo_drainage_cannula_insertion_site, complication_stroke) %>%
  tally() %>%
  group_by(ecmo_drainage_cannula_insertion_site) %>%
  mutate(percent = prop.table(n)*100,
         cell = paste(n, ' (', roundz(percent, digits=0), ')', sep='')) %>%
  ungroup() %>%
  mutate(ecmo_drainage_cannula_insertion_site = ifelse(ecmo_drainage_cannula_insertion_site=='', 'Missing', ecmo_drainage_cannula_insertion_site ),
         complication_stroke = ifelse(is.na(complication_stroke), 'Missing', as.character(complication_stroke))) %>%
  select(ecmo_drainage_cannula_insertion_site, complication_stroke, cell) %>%
  spread(complication_stroke, cell)
#
table_header <- data.frame(
  col_keys = c( 'ecmo_drainage_cannula_insertion_site', 'Missing', 'No', 'Yes' ),
  h1 = c('', rep("Stroke complication", 3)),
  h2 = c('Drainage cannula insertion site', 'Missing', 'No', 'Yes'),
  stringsAsFactors = FALSE )
#
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>%
  autofit()
ftab
```
The tables show the number of patients and the row percentage.

\newpage
## Association between a single or double lumen cannula and stroke/ICH/death

The table below shows the association between cannula type and stroke complication.

```{r, results='asis'}
tab = group_by(ecmo_patients, cannula_lumen, complication_stroke) %>%
  filter() %>%
  tally() %>%
  group_by(cannula_lumen) %>%
  mutate(percent = prop.table(n)*100,
         cell = paste(n, ' (', roundz(percent, digits=0), ')', sep='')) %>%
  ungroup() %>%
  mutate(cannula_lumen = ifelse(cannula_lumen=='', 'Missing', cannula_lumen ),
         complication_stroke = ifelse(is.na(complication_stroke), 'Missing', as.character(complication_stroke))) %>%
  select(cannula_lumen, complication_stroke, cell) %>%
  spread(complication_stroke, cell)
#
table_header <- data.frame(
  col_keys = c( 'cannula_lumen', 'Missing', 'No', 'Yes' ),
  h1 = c('', rep("Stroke complication", 3)),
  h2 = c('Type of cannula', 'Missing', 'No', 'Yes'),
  stringsAsFactors = FALSE )
#
ftab = flextable(tab) %>%
  set_header_df(mapping = table_header, key = "col_keys" ) %>% # add header
  merge_h(i=1, part = "header") %>% # merge "To" column headers
  theme_box() %>%
  autofit()
ftab
```


The tables show the number of patients and the row percentage.
